requirements.txt
# # --------- pytorch --------- #
https://download.pytorch.org/whl/cu118/torch-2.1.2%2Bcu118-cp310-cp310-linux_x86_64.whl
https://download.pytorch.org/whl/cu118/torchaudio-2.1.2%2Bcu118-cp310-cp310-linux_x86_64.whl
https://download.pytorch.org/whl/cu118/torchvision-0.16.2%2Bcu118-cp310-cp310-linux_x86_64.whl

lightning[extra]==2.1.2
torchmetrics==0.11.4

# # # --------- model zoo -------#
transformers==4.39.1
datasets==2.15.0
timm==0.9.2

# # ------- tensor utils ------#
einops==0.6.1

# # ------- model training ------#
peft==0.10.0
bitsandbytes==0.43.0
accelerate==0.28.0
trl==0.8.1
https://github.com/Dao-AILab/flash-attention/releases/download/v2.5.6/flash_attn-2.5.6+cu118torch2.1cxx11abiFALSE-cp310-cp310-linux_x86_64.whl

# # --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# # --------- loggers --------- #
wandb==0.15.4
mlflow==2.4.1       # install this later with pip install as it is failing with devcontainer build as its trying to uninstall blinker==1.4 and install 1.6 instead
aim==3.17.5         # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550
tensorboard==2.16.2
tensorboardX==2.6.2.2
neptune==1.3.1
comet-ml==3.33.5

# # --------- deployment ------ #
fastapi[all]==0.110.0
vllm==0.4.0.post1
openai==1.16.1

# # --------- others --------- #
pyrootutils==1.0.4      # standardizing the project root setup
pre-commit==3.3.3       # hooks for applying linters on commit
rich==13.4.2            # beautiful text formatting in terminal
pytest==7.3.2           # testing
black==23.3.0           # code formatting
twine==4.0.2            # for uploading to pypi
sh==2.0.4               # for running bash commands in some tests (linux/macos only)
